Up and running with the mechanics of Dialogs
